Rename matasano-converter files to match the name of the class they describe.  Start sketch of the getBase64 function
